Split admin review events report subfeatures
